# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "HERA 2.0 Mobile API",
    "version": "2.0.0",
    "description": "API untuk integrasi aplikasi mobile HERA (Hydrology Environmental Real-time Analytics) dengan sistem monitoring kualitas air berbasis IoT. Semua endpoint yang dilindungi memerlukan Bearer Token dari Sanctum.\n\n## Cara Penggunaan\n1. Dapatkan token via `POST /api/mobile/login` atau `POST /api/mobile/register`\n2. Klik tombol **Authorize** di atas dan masukkan token\n3. Coba setiap endpoint langsung dari halaman ini",
    "contact": {
      "name": "Tim HERA 2.0 — Politeknik Negeri Lampung",
      "email": "admin@hera.ac.id"
    }
  },
  "servers": [
    {
      "url": "http://103.217.145.187:8000",
      "description": "HERA 2.0 VPS Production"
    },
    {
      "url": "http://127.0.0.1:8000",
      "description": "Lokal Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "Masukkan token dari endpoint /api/mobile/login atau /api/mobile/register"
      }
    },
    "schemas": {
...
      "FieldTestItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 7 },
          "user_id": { "type": "integer", "example": 3 },
          "latitude": { "type": "number", "format": "float", "example": -6.967585 },
          "longitude": { "type": "number", "format": "float", "example": 107.659063 },
          "altitude": { "type": "number", "format": "float", "example": 768.5, "nullable": true },
          "ph": { "type": "number", "format": "float", "example": 7.5 },
          "tds": { "type": "number", "format": "float", "example": 300 },
          "ec": { "type": "number", "format": "float", "example": 450 },
          "suhu_air": { "type": "number", "format": "float", "example": 28.5 },
          "suhu_lingkungan": { "type": "number", "format": "float", "example": 30.0 },
          "kelembapan": { "type": "number", "format": "float", "example": 75.0 },
          "tegangan": { "type": "number", "format": "float", "example": 3.8 },
          "cr_estimated": { "type": "number", "format": "float", "example": 0.01545, "description": "Estimasi kadar Chromium (mg/L) dari model AI" },
          "created_at": { "type": "string", "format": "date-time", "example": "2026-04-16T06:30:18.000000Z" },
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "integer", "example": 3 },
              "name": { "type": "string", "example": "Petugas Monitoring" }
            }
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/mobile/testing/location": {
      "post": {
        "tags": ["Field Test"],
        "summary": "Kirim data pengujian lapangan",
        "description": "Menyimpan data hasil pengujian lapangan dari mobile. GPS + data sensor disimpan ke database. Sistem secara otomatis menghitung estimasi kadar Chromium menggunakan model AI FastAPI.\n\n**Altitude** (ketinggian) bersifat opsional — jika perangkat mendukung GPS altimeter, nilai ini akan tersimpan dan ditampilkan di peta web dashboard.",
        "operationId": "fieldTestStore",
        "security": [{ "sanctum": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["latitude", "longitude"],
                "type": "object",
                "properties": {
                  "latitude": { "type": "number", "format": "float", "example": -6.967585 },
                  "longitude": { "type": "number", "format": "float", "example": 107.659063 },
                  "altitude": { "type": "number", "format": "float", "example": 768.5, "description": "Ketinggian lokasi (meter dpl, opsional)" },
                  "suhu_air": { "type": "number", "format": "float", "example": 28.5 },
                  "suhu_lingkungan": { "type": "number", "format": "float", "example": 30.0 },
                  "kelembapan": { "type": "number", "format": "float", "example": 75.0 },
                  "ec": { "type": "number", "format": "float", "example": 450.0, "description": "Electrical Conductivity (mS/cm)" },
                  "tds": { "type": "number", "format": "float", "example": 300.0, "description": "Total Dissolved Solids (ppm)" },
                  "ph": { "type": "number", "format": "float", "example": 7.5 },
                  "tegangan": { "type": "number", "format": "float", "example": 3.8, "description": "Tegangan baterai sensor (V)" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Data pengujian berhasil disimpan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Data pengujian terkalibrasi berhasil terekam!" },
                    "data": { "$ref": "#/components/schemas/FieldTestItem" }
                  }
                }
              }
            }
          },
          "401": { "description": "Token tidak valid" },
          "422": { "description": "Validasi gagal — latitude/longitude wajib diisi" }
        }
      }
    },
    "/api/mobile/testing/history": {
      "get": {
        "tags": ["Field Test"],
        "summary": "Riwayat pengujian lapangan",
        "description": "Mengambil riwayat data pengujian lapangan semua petugas dengan cursor-based pagination.\n\n**Cara paginasi:**\n1. Panggil tanpa parameter `cursor` untuk halaman pertama\n2. Gunakan nilai `next_cursor` dari response sebagai parameter `cursor` untuk halaman berikutnya\n3. Jika `next_cursor` bernilai `null`, artinya sudah halaman terakhir",
        "operationId": "fieldTestHistory",
        "security": [{ "sanctum": [] }],
        "parameters": [
          { "name": "limit", "in": "query", "description": "Jumlah item per halaman (default: 20)", "schema": { "type": "integer", "example": 20 } },
          { "name": "cursor", "in": "query", "description": "ISO datetime dari `next_cursor` response sebelumnya. Kosongkan untuk halaman pertama.", "schema": { "type": "string", "example": "2026-04-15T13:16:28+07:00" } }
        ],
        "responses": {
          "200": {
            "description": "Riwayat pengujian lapangan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Success retrieve history" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "data": { "type": "array", "items": { "$ref": "#/components/schemas/FieldTestItem" } },
                        "next_cursor": { "type": "string", "nullable": true, "example": "2026-04-15T13:16:28+07:00", "description": "Kirim nilai ini sebagai `cursor` untuk halaman berikutnya. Null jika sudah halaman terakhir." }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Token tidak valid" }
        }
      }
    }
  }
}